Use a single BookEditDetailsViewController for editing & searching (#68)
* Initial draft: Single view controller for book details

* Code cleanup

* fit-n-finish
