+++ GrailDiary/SavingTextEditViewController.swift
@@ -176,7 +176,8 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
   }
 
   func editBookDetails(book: AugmentedBook) {
-    let bookViewController = BookEditDetailsViewController(book: book, coverImage: coverImage)
+    guard let apiKey = ApiKey.googleBooks else { return }
+    let bookViewController = BookEditDetailsViewController(apiKey: apiKey, book: book, coverImage: coverImage, showSkipButton: false)
     bookViewController.delegate = self
     bookViewController.title = "Edit Book Details"
     let navigationController = UINavigationController(rootViewController: bookViewController)
@@ -185,7 +186,7 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
   }
 
   func insertBookDetails(apiKey: String) {
-    let bookViewController = BookSearchViewController(apiKey: apiKey, showSkipButton: false)
+    let bookViewController = BookEditDetailsViewController(apiKey: apiKey, showSkipButton: false)
     bookViewController.delegate = self
     bookViewController.title = "Insert Book Details"
     let navigationController = UINavigationController(rootViewController: bookViewController)
@@ -320,9 +321,9 @@ final class SavingTextEditViewController: UIViewController, TextEditViewControll
     }
   }
 
-  func textEditViewController(_ viewController: TextEditViewController, didAttach book: Book) {
+  func textEditViewController(_ viewController: TextEditViewController, didAttach book: AugmentedBook) {
     Logger.shared.info("Attaching book: \(book.title)")
-    note.metadata.book = AugmentedBook(book)
+    note.metadata.book = book
     note.metadata.modifiedTimestamp = Date()
     tryUpdateNote(note)
   }
@@ -405,8 +406,8 @@ extension SavingTextEditViewController: BookHeaderDelegate {
   }
 }
 
-extension SavingTextEditViewController: BookSearchViewControllerDelegate {
-  public func bookSearchViewController(_ viewController: BookSearchViewController, didSelect book: Book, coverImage: UIImage?) {
+extension SavingTextEditViewController: BookEditDetailsViewControllerDelegate {
+  public func bookSearchViewController(_ viewController: BookEditDetailsViewController, didSelect book: AugmentedBook, coverImage: UIImage?) {
     if let image = coverImage, let imageData = image.jpegData(compressionQuality: 0.8) {
       do {
         _ = try storeImageData(imageData, type: .jpeg, key: type(of: noteStorage).coverImageKey)
@@ -415,32 +416,32 @@ extension SavingTextEditViewController: BookSearchViewControllerDelegate {
       }
     }
     textEditViewController(textEditViewController, didAttach: book)
-    textEditViewController.extendedNavigationHeaderView = BookHeader(book: AugmentedBook(book), coverImage: coverImage)
-    dismiss(animated: true, completion: nil)
-  }
-
-  public func bookSearchViewControllerDidSkip(_ viewController: BookSearchViewController) {
-    // NOTHING
-  }
-
-  public func bookSearchViewControllerDidCancel(_ viewController: BookSearchViewController) {
-    dismiss(animated: true, completion: nil)
-  }
-}
-
-extension SavingTextEditViewController: BookEditDetailsViewControllerDelegate {
-  func bookEditDetailsViewControllerDidCancel(_ viewController: BookEditDetailsViewController) {
-    dismiss(animated: true, completion: nil)
-  }
-
-  func bookEditDetailsViewController(_ viewController: BookEditDetailsViewController, didFinishEditing book: AugmentedBook, coverImage: UIImage?) {
-    Logger.shared.info("Attaching book: \(book.title)")
-    self.coverImage = coverImage
-    note.metadata.book = book
-    note.metadata.modifiedTimestamp = Date()
-    tryUpdateNote(note)
-    let coverImage = coverImageCache.coverImage(bookID: noteIdentifier, maxSize: 250)
     textEditViewController.extendedNavigationHeaderView = BookHeader(book: book, coverImage: coverImage)
     dismiss(animated: true, completion: nil)
   }
+
+  public func bookSearchViewControllerDidSkip(_ viewController: BookEditDetailsViewController) {
+    // NOTHING
+  }
+
+  public func bookSearchViewControllerDidCancel(_ viewController: BookEditDetailsViewController) {
+    dismiss(animated: true, completion: nil)
+  }
 }
+
+//extension SavingTextEditViewController: BookEditDetailsViewControllerDelegate {
+//  func bookEditDetailsViewControllerDidCancel(_ viewController: BookEditDetailsViewController) {
+//    dismiss(animated: true, completion: nil)
+//  }
+//
+//  func bookEditDetailsViewController(_ viewController: BookEditDetailsViewController, didFinishEditing book: AugmentedBook, coverImage: UIImage?) {
+//    Logger.shared.info("Attaching book: \(book.title)")
+//    self.coverImage = coverImage
+//    note.metadata.book = book
+//    note.metadata.modifiedTimestamp = Date()
+//    tryUpdateNote(note)
+//    let coverImage = coverImageCache.coverImage(bookID: noteIdentifier, maxSize: 250)
+//    textEditViewController.extendedNavigationHeaderView = BookHeader(book: book, coverImage: coverImage)
+//    dismiss(animated: true, completion: nil)
+//  }
+//}
